refactor: 💡 remove route action usage for scopes
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-13755
