+++ ui/admin/app/components/form/onboarding/index.hbs
@@ -7,6 +7,7 @@
   class='full-width'
   @cancel={{@cancel}}
   @onSubmit={{this.submit}}
+  @disabled={{this.isSaving}}
   as |form|
 >
   <div class='create-resources-form'>
