+++ ui/admin/app/controllers/scopes/scope/authenticate/method/oidc.js
@@ -3,10 +3,12 @@
  * SPDX-License-Identifier: BUSL-1.1
  */
 
-import Controller from '@ember/controller';
+import Controller, { inject as controller } from '@ember/controller';
 import { inject as service } from '@ember/service';
 
 export default class ScopesScopeAuthenticateMethodOidcController extends Controller {
+  @controller('scopes/scope/authenticate/method/index') method;
+
   // =services
 
   @service session;
