+++ ui/admin/app/controllers/scopes/scope/edit.js
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: BUSL-1.1
+ */
+
+import Controller, { inject as controller } from '@ember/controller';
+
+export default class ScopesScopeEditController extends Controller {
+  @controller('scopes/scope/index') scopes;
+}
