+++ ui/admin/app/controllers/scopes/scope/scopes/new.js
@@ -1,0 +1,5 @@
+import Controller, { inject as controller } from '@ember/controller';
+
+export default class ScopesScopeScopesNewController extends Controller {
+  @controller('scopes/scope/index') scopes;
+}
