+++ ui/admin/app/routes/onboarding.js
@@ -5,9 +5,6 @@
 
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
-import { action } from '@ember/object';
-import { loading } from 'ember-loading';
-import { notifyError } from 'core/decorators/notify';
 import { TYPE_TARGET_TCP } from 'api/models/target';
 
 export default class OnboardingRoute extends Route {
@@ -16,7 +13,6 @@ export default class OnboardingRoute extends Route {
   @service store;
   @service session;
   @service router;
-  @service intl;
 
   // =methods
 
@@ -58,96 +54,5 @@ export default class OnboardingRoute extends Route {
         name: 'test_target_role',
       }),
     };
-  }
-
-  // =actions
-  @action
-  @loading
-  @notifyError(function () {
-    return this.intl.t('errors.onboarding-failed.description');
-  })
-  async createResources(targetAddress, targetPort) {
-    const { org, project, target, role } = this.currentModel;
-
-    await this.createOrg(org);
-    await this.createProject(org, project);
-    await this.createTarget(project, target, targetAddress, targetPort);
-    await this.createGrants(org, project, target, role);
-
-    // Redirect to success
-    this.router.transitionTo('onboarding.success');
-  }
-
-  /**
-   * Creates Org, if fails, redirects to Orgs empty list
-   * @param {Object} org
-   */
-  async createOrg(org) {
-    try {
-      await org.save();
-    } catch (error) {
-      // Redirects to Orgs screen
-      this.router.transitionTo('scopes.scope', 'global');
-      throw new Error(error);
-    }
-  }
-
-  /**
-   * Creates Project, if fails redirects to Projects from Org empty list.
-   * @param {Object} org
-   * @param {Object} project
-   */
-  async createProject(org, project) {
-    project.scopeID = org.id;
-
-    try {
-      await project.save();
-    } catch (error) {
-      // Redirects to Project screen
-      this.router.transitionTo('scopes.scope.scopes', project.scopeID);
-      throw new Error(error);
-    }
-  }
-
-  /**
-   * Creates a Target, if fails redirects to Targets empyt list.
-   * @param {Object} project
-   * @param {Object} target
-   * @param {Object} targetAddress
-   * @param {Object} targetPort
-   */
-  async createTarget(project, target, targetAddress, targetPort) {
-    target.scopeID = project.id;
-    target.address = targetAddress;
-    target.default_port = targetPort;
-
-    try {
-      await target.save();
-    } catch (error) {
-      // Redirect to targets global
-      this.router.replaceWith('scopes.scope.targets', project.scopeID);
-      throw new Error(error);
-    }
-  }
-
-  async createGrants(org, project, target, role) {
-    role.scopeID = org.id;
-    role.grant_scope_id = project.id;
-    try {
-      await role.save();
-      await role.saveGrantStrings([
-        `type=target;actions=list`,
-        `ids=${target.id};actions=authorize-session`,
-      ]);
-    } catch (error) {
-      // Redirect to targets that belong to scope
-      this.router.replaceWith('scopes.scope.targets', project.scopeID);
-      throw new Error(error);
-    }
-  }
-
-  @action
-  async doLater() {
-    this.router.transitionTo('index');
   }
 }
