+++ ui/admin/app/templates/onboarding/success.hbs
@@ -86,7 +86,7 @@
       <Hds::Button
         @text={{t 'actions.done'}}
         data-test-onboarding-done-button
-        {{on 'click' (route-action 'showTargetList' @model)}}
+        {{on 'click' (fn this.showTargetList @model)}}
       />
     </:body>
   </Rose::Frame>
