+++ ui/admin/app/templates/scopes/scope/authenticate/method/index.hbs
@@ -5,6 +5,6 @@
 
 <Form::Authenticate
   @model={{@model}}
-  @submit={{route-action 'authenticate'}}
+  @submit={{fn this.authenticate @model}}
   @disabled={{(is-loading)}}
 />
