+++ ui/admin/app/templates/scopes/scope/authenticate/method/oidc.hbs
@@ -12,7 +12,7 @@
     <p>{{t 'resources.auth-method.messages.pending.description'}}</p>
     <p>
       {{t 'resources.auth-method.questions.no-see-window'}}
-      <a href='#' role='button' {{on 'click' (route-action 'authenticate')}}>
+      <a href='#' role='button' {{on 'click' this.method.authenticate}}>
         {{t 'actions.retry'}}
       </a>
     </p>
