+++ ui/admin/app/templates/scopes/scope/scopes/new.hbs
@@ -41,8 +41,8 @@
   <page.body>
     <Form::Scope
       @model={{@model}}
-      @submit={{route-action 'save' @model}}
-      @cancel={{route-action 'cancel' @model}}
+      @submit={{fn this.scopes.save @model}}
+      @cancel={{fn this.scopes.cancel @model}}
     />
   </page.body>
